found way to send values in E correctly. something is wrong with G. For plot I have to let v run up to 3*W to cover full image...

+++ src/delaunay_triangulation.cpp
@@ -155,7 +155,7 @@ void delaunay_triangulation(cv::Mat &S, int H, int W, cv::Mat &G, cv::Mat &T, cv
 	
 
 	k = 0;
-	E = cv::Mat(2 * out.numberofedges, 1, CV_16U);
+	E = cv::Mat(2 * out.numberofedges, 1, CV_32S);
 	for (int i = 0; i < out.numberofedges; ++i)
 	{
 		// For plotting triangle edges
